pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-chatbot"
version = "0.1.0"
description = "Simple RAG chatbot practice project for internship preparation."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.5.5",
    "PyMuPDF>=1.24.7",
    "python-dotenv>=1.0.1",
    "requests>=2.32.0",
    "sentence-transformers>=3.0.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
